+++ src/Users.js
@@ -33,12 +33,12 @@ const filterConfig = [
       {
         name: 'inactive',
         cql: 'false',
-        displayName: <FormattedMessage id="ui-users.inactiveUserStatus" />,
+        displayName: <FormattedMessage id="ui-users.inactive" />,
       },
       {
         name: 'active',
         cql: 'true',
-        displayName: <FormattedMessage id="ui-users.activeUserStatus" />,
+        displayName: <FormattedMessage id="ui-users.active" />,
       },
     ],
   },
@@ -72,7 +72,7 @@ class Users extends React.Component {
           query: makeQueryFunction(
             'cql.allRecords=1',
             // TODO: Refactor/remove this after work on FOLIO-2066 and RMB-385 is done
-            (parsedQuery, props, localProps) => localProps.query.query.split(' ').map(query => compileQuery({ query })).join(' and '),
+            (parsedQuery, props, localProps) => localProps.query.query.trim().split(/\s+/).map(query => compileQuery({ query })).join(' and '),
             {
               // the keys in this object must match those passed to
               // SearchAndSort's columnMapping prop