+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Remove unnecessary settings. Refs UIORG-150.
 * Handle "Active/Inactive" hard-code issue for the search filter. Fixes UIU-894.
+* Better search string construction won't accidentally match `*`. Fixes UIU-1086.
 
 ## [2.23.0](https://github.com/folio-org/ui-users/tree/v2.23.0) (2019-06-12)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v2.22.0...v2.23.0)