+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Remove unnecessary settings. Refs UIORG-150.
 * Handle "Active/Inactive" hard-code issue for the search filter. Fixes UIU-894.
+* Better search string construction won't accidentally match `*`. Fixes UIU-1086.
 * Fix sort patron blocks. Fixes UIU-868.
 * Fix Payment Status from "Charge and pay now". Fixes UIU-931.
 * Fix UX Consistency owners. Fixes UIU-1050.