+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Fix sort patron blocks. Fixes UIU-868.
 * Fix Payment Status from "Charge and pay now". Fixes UIU-931.
 * Fix UX Consistency owners. Fixes UIU-1050.
+* Handle address type validation issue after adding permission to user. Fixes UIU-912.
 
 ## [2.23.0](https://github.com/folio-org/ui-users/tree/v2.23.0) (2019-06-12)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v2.22.0...v2.23.0)