+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Add missing permissions. Fixes UIU-1088.
 * Accomodate changes in the shape of `/loan-storage/loan-history` data. Refs UIU-1092.
 * Fix expiration date for sponsor and proxy. Fixes UIU-924.
+* Update tests to cover changes in due date modal in stripes-smart-components. Refs UIU-1070. 
 
 ## [2.23.0](https://github.com/folio-org/ui-users/tree/v2.23.0) (2019-06-12)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v2.22.0...v2.23.0)