+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Change history for ui-users
 
-## 2.24.0 (IN PROGRESS)
+## [2.24.0](https://github.com/folio-org/ui-users/tree/v2.24.0) (2019-07-24)
+[Full Changelog](https://github.com/folio-org/ui-users/compare/v2.23.0...v2.24.0)
 
+* Allow some loan-renewal failures to be overridden. Refs UIU-1017.
 * Remove unnecessary settings. Refs UIORG-150.
 * Handle "Active/Inactive" hard-code issue for the search filter. Fixes UIU-894.
 * Better search string construction won't accidentally match `*`. Fixes UIU-1086.
@@ -18,6 +20,7 @@
 * Create BigTest Payment methods. Refs UIU-998.
 * Create BigTest Refund reasons. Refs UIU-999.
 * Create BigTest Waive reasons. Refs UIU-1001.
+* Update `login` API to `v6.0`. Refs UIU-1099.
 
 ## [2.23.0](https://github.com/folio-org/ui-users/tree/v2.23.0) (2019-06-12)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v2.22.0...v2.23.0)